refactor: 2단계 - 로또, 리팩토링

+++ src/main/java/lotto/domain/PrizeMoney.java
@@ -2,13 +2,13 @@ package lotto.domain;
 
 public class PrizeMoney {
 
-	public final int prizeMoney;
+	public final long prizeMoney;
 
-	public PrizeMoney(int prizeMoney) {
+	public PrizeMoney(long prizeMoney) {
 		this.prizeMoney = prizeMoney;
 	}
 
-	public int getPrizeMoney() {
+	public long getPrizeMoney() {
 		return this.prizeMoney;
 	}
 }
